Fixed bug in subscription controller

+++ DAL/Repo/IProviderRepository.cs
@@ -6,7 +6,7 @@ namespace DAL.Repo
     public interface IProviderRepository
     {
         void Add(Provider obj);
-        List<Provider> GetProviders(int amount, int duration, string type);
+        List<Provider> GetProviders(decimal amount, int duration, string type);
         List<Provider> GetAllProviders();
         Provider GetById(int id);
         User GetUserById(string id);
